+++ src/test/java/org/mmarini/wheelly/model/GridScannerMapEchoTest.java
@@ -43,7 +43,7 @@ import java.util.stream.Stream;
 import static java.lang.Math.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
-import static org.mmarini.wheelly.model.AbstractScannerMap.*;
+import static org.mmarini.wheelly.model.GridScannerMap.*;
 import static org.mmarini.wheelly.model.GridScannerMap.snapToGrid;
 import static org.mmarini.wheelly.model.Utils.linear;